requirements.txt
absl-py==2.3.0
accelerate==1.7.0
ai-edge-litert==1.3.0
ai-edge-quantizer==0.2.1
ai-edge-torch==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
astunparse==1.6.3
attrs==25.3.0
backports.strenum==1.2.8
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.1
frozenlist==1.6.2
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.72.1
h5py==3.13.0
hf-xet==1.1.3
huggingface-hub==0.32.4
idna==3.10
immutabledict==4.2.1
iniconfig==2.1.0
itsdangerous==2.2.0
jax==0.6.2
jax-cuda12-pjrt==0.6.2
jax-cuda12-plugin==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
kagglehub==0.3.12
keras==3.10.0
keras-nightly==3.10.0.dev2025060403
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.4
multipledispatch==1.0.0
multiprocess==0.70.16
namex==0.1.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
peft==0.15.2
pillow==11.2.1
pluggy==1.6.0
propcache==0.3.1
protobuf==4.25.8
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
setuptools==78.1.1
six==1.17.0
sounddevice==0.5.2
sympy==1.14.0
tabulate==0.9.0
tb-nightly==2.19.0a20250218
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
tf_keras==2.19.0
tf_nightly==2.20.0.dev20250515
tokenizers==0.21.1
torch==2.7.1
torch_xla2==0.0.1.dev202412041639
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
triton==3.3.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
waitress==3.0.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
